Add support for site specific assets
Here we are. In an attempt to fix #13 there is now
a new task: `gonz.serve` which will serve the built
static site. The reason we need this now is because
the simplest way to refer to an asset is by using a
non-relative path, like /assets/images/pangolin.png.
Now browsing the site locally that will not work so we
need to add a web server which will map "/" to whatever
we chose. Oh well.

This change did bring some simplifications to other parts.
However you will still not be able to just render the markdown
and see an image for example, because again the path will
start at the root.